requirements.txt
Django~=3.1.0
django-widget-tweaks~=1.4.8
django-allauth~=0.41.0
Pillow~=8.1.0